pyproject.toml
[project]
name = "zahnfunnel"
version = "0.1.0"
description = "Zahnarzt-Lead-Funnel: Formular- und WhatsApp-Eingangspfade mit AI-personalisierten Antworten"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "pydantic>=2.9",
    "pydantic-settings>=2.6",
    "psycopg[binary,pool]>=3.2",
    "httpx>=0.27",
    "anthropic>=0.39",
    "google-api-python-client>=2.149",
    "google-auth>=2.35",
    "google-auth-oauthlib>=1.2",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.3",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]
exclude = ["db*", "n8n*", "secrets*", "tests*"]
